Fixed more errors.
