+++ coin-alert/src/app/api/checkPriceAlerts/route.ts
@@ -1,127 +1,7 @@
-import { connection } from "@/app/lib/connection";
-import { PriceData, tokenConverter } from "@/app/lib/firebase/tokenUtils";
 import { getAllUsers, SirenUser } from "@/app/lib/firebase/userUtils";
-import { blockchainTaskQueue } from "@/app/lib/taskQueue";
-import { TokenAccountData } from "@/app/lib/utils/solanaUtils";
-import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
-import { PublicKey } from "@solana/web3.js";
-import { doc, getDoc } from "firebase/firestore";
-import { db } from "../../lib/firebase/firebase";
+import { AlarmConfig, calculatePriceChange, getAlarmConfig, getLastHourPrices, getTokensFromBlockchain, NotificationReturn } from "@/app/lib/utils/priceAlertHelper";
 import { sendNotification } from "../../lib/sendNotifications"; // Push notification logic
 
-
-export interface AlarmConfig {
-  standardAlarmPercentage: number;
-  criticalAlarmPercentage: number;
-}
-
-export type AlarmType = "normal" | "critical" | null 
-
-interface NotificationReturn {
-  userId: string,
-  token: string
-  priceChange: number,
-  alertType: AlarmType
-  minutes: number
-  alarmedConfig: AlarmConfig | null
-  percentageBreached: number
-}
-
-const ALARM_CONFIGS = new Map<number, AlarmConfig>([
-  [1, { standardAlarmPercentage: 50, criticalAlarmPercentage: 80}],
-  [7, { standardAlarmPercentage: 60, criticalAlarmPercentage: 90}],
-  [15, {standardAlarmPercentage: 70, criticalAlarmPercentage: 100}],
-  [30, {standardAlarmPercentage: 80, criticalAlarmPercentage: 120}],
-  [60, {standardAlarmPercentage: 90, criticalAlarmPercentage: 175}]
-])
-
-const QUIETER_ALARM_CONFIGS = new Map<number, AlarmConfig>();
-const NOISIER_ALARM_CONFIGS = new Map<number, AlarmConfig>();
-
-ALARM_CONFIGS.forEach((config, key) => {
-  QUIETER_ALARM_CONFIGS.set(key, {
-    standardAlarmPercentage: config.standardAlarmPercentage * 2,
-    criticalAlarmPercentage: config.criticalAlarmPercentage * 2
-  });
-});
-
-
-ALARM_CONFIGS.forEach((config, key) => {
-  NOISIER_ALARM_CONFIGS.set(key, {
-    standardAlarmPercentage: config.standardAlarmPercentage / 4,
-    criticalAlarmPercentage: config.criticalAlarmPercentage / 4
-  });
-});
-
-console.log("Noisey configs: " + NOISIER_ALARM_CONFIGS)
-
-
-
-async function getLastHourPrices(token: string): Promise<PriceData[]> {
-  try {
-    console.log("Getting last hour prices for token: " + token);
-
-    const oneHourAgo = Date.now() - 60 * 60 * 1000; // 1 hour ago in milliseconds
-    console.log("One hour ago (ms): " + oneHourAgo);
-
-    const tokenDocRef = doc(db, "uniqueTokens", token).withConverter(tokenConverter);
-    const tokenSnapshot = await getDoc(tokenDocRef);
-
-    if (!tokenSnapshot.exists()) {
-      console.warn(`⚠️ No document found for token: ${token}`);
-      return [];
-    }
-
-    const tokenData = tokenSnapshot.data();
-
-    if (!tokenData?.prices || !Array.isArray(tokenData.prices)) {
-      console.warn(`⚠️ No price history found for token: ${token}`);
-      return [];
-    }
-
-    // 🔹 Filter prices to only include last 60 minutes
-    const lastHourPrices = tokenData.prices
-      .filter((entry: PriceData) => entry.timestamp > oneHourAgo)
-      .sort((a, b) => b.timestamp - a.timestamp); // Sort by newest first
-
-    console.log(`✅ Found ${lastHourPrices.length} price entries for token: ${token}`);
-    return lastHourPrices;
-  } catch (error) {
-    console.error(`❌ Error fetching prices for ${token}:`, error);
-    return [];
-  }
-}
-
-
-function getAlarmConfig(){
-  // console.log("Using alarm config: ")
-  // console.log(NOISIER_ALARM_CONFIGS)
-  return NOISIER_ALARM_CONFIGS
-}
-
-
-// 🔹 Placeholder: Fetch Tokens Owned by User (Implement This Later)
-async function getTokensFromBlockchain(walletAddress: string): Promise<string[]> {
-  const publicKey = new PublicKey(walletAddress);
-  const tokenAccountsForAddress = await blockchainTaskQueue.addTask(() => connection.getParsedTokenAccountsByOwner(publicKey, { programId: TOKEN_PROGRAM_ID })) 
-  const tokensHeldByAddress = tokenAccountsForAddress.value.filter((val) => {
-    const tokenAccountData: TokenAccountData = val.account.data.parsed
-    if ((tokenAccountData.info.tokenAmount.uiAmount || 0) > 0) {
-      return true
-    } else {
-      return false
-    }
-  }).map((val) => {
-    const tokenAccountData: TokenAccountData = val.account.data.parsed
-    return tokenAccountData.info.mint
-  })
-  return tokensHeldByAddress
-}
-
-// 🔹 Check Price Change Percentage
-function calculatePriceChange(oldPrice: number, newPrice: number): number {
-  return ((newPrice - oldPrice) / oldPrice) * 100;
-}
 
 // 🔹 Main API Function
 export async function GET(req: Request) {
