+++ coin-alert/src/app/lib/connection.ts
@@ -1,3 +1,7 @@
 import { Connection } from "@solana/web3.js";
 
-export const connection = new Connection(process.env.RPC_ENDPOINT || "")
+if(!process.env.RPC_ENDPOINT?.length){
+    throw new Error("ERROR: Cannot conncet to RPC node.")
+}
+
+export const connection = new Connection(process.env.RPC_ENDPOINT)
